ca_ci.cpp: supplement to 9a9372931abf129833100ee21a8be1216a8b67b2

+++ common/ca_ci.cpp
@@ -56,7 +56,7 @@ static cCA* pcCAInstance = NULL;
 /* nur diese Message wird vom CI aus neutrinoMessages.h benutzt */
 /* für den CamMsgHandler, darum hier einfach mal definiert */
 /* die Feinheiten werden ja in CA_MESSAGE verpackt */
-uint32_t EVT_CA_MESSAGE = 0x80000000 + 60;
+uintptr_t EVT_CA_MESSAGE = 0x80000000 + 60;
 
 static cs_messenger cam_messenger = NULL;
